One global commit

+++ src/Application/Features/Comments/Commands/DeleteComment.cs
@@ -2,6 +2,7 @@
 
 using Application.Core;
 using Infrastructure;
+using Microsoft.EntityFrameworkCore;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -16,10 +17,18 @@ public class DeleteComment
     {
         public async Task<Result<bool>> Handle(Command request, CancellationToken cancellationToken)
         {
-            Comment? comment = await context.Comments.FindAsync(request.CommentId, cancellationToken);
+            Comment? comment = await context.Comments.Include(a => a.Replies).FirstOrDefaultAsync(c => c.Id == request.CommentId, cancellationToken);
             if (comment == null) return Result<bool>.Failure("Comment was not found", 404);
 
-            context.Comments.Remove(comment);
+            if (comment.Replies.Count == 0)
+            {
+                context.Comments.Remove(comment);
+                return Result<bool>.Success(true);
+            }
+
+            comment.IsDeleted = true;
+
+            context.Update(comment);
             return Result<bool>.Success(true);
         }
     }
